usb: xhci: pass nonblock argument of usb_int_msg to xhci-ring
For some misbehaved keyboards, they only reply NAKs when no
key pressed. Thus causing usb_int_msg always block 5s when
no inputs, which reduces the responsiveness of other app.
It's not unnecessary to wait for so long, fix this by passing
through the nonblock argument of usb_int_msg to xhci, and use
a smaller timeout for that.

Change-Id: I1bddea617e35a087b74447dc4b6ac1d39f50f0fe