hdmi: set xsize & ysize even if hdmi or mipi is NOT connect
GOP(graphic out procotol) use xsize and ysize to initialize its
height and width in uboot; if xsize & ysize were set to 0, height
and width in gopobj->info would be set to 0.

Grub get height and width through GOP, when they were all set to 0, it
will be overflow during the calculation of the new width and height,
and it malloc a huge memory space in grub_virtual_screen_setup, which
cause memory space was exhausted.

Change-Id: I21e7c672ec8e01229f9fc80d3187f7a97c019a41